+++ CHANGELOG.md
@@ -7,11 +7,13 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.3.5 - Unreleased
 
+### Changed
+* kwcoco will now use ujson to read json files by default if it is available.
+* Added runtime deprecation noticed to deprecated functions
+* Doctests now generate images for readthedocs.
 
-## Version 0.3.4 - Released 2022-08-06
 
-
-## Version 0.3.3 - Released 2022-08-06
+## Version 0.3.3 - Released 2022-07-26
 
 ### Added
 * Ability to "undo warps" from delayed ops.